requirements.txt
Pillow
Gradio
requests
paddlepaddle
paddleocr